v2.8
Waiting for a new response no longer 'Discourages' the AI and it should
now maintain the topic, so the AI can now give "multiple responses".
Also fixed a bug with the Discouraging that would cause the pre/pro word
files to lose entries whose frequencies were reduced to 0... this lead
to the AI repeating a single word when it would attempt to generate a
response, since it would no longer find any words it learned to go
before/after the identified topic. Another bug that would cause pre/pro
word files to be needlessly created was also fixed.