v2.9
Created a screen with instructions for enabling the 'Storage' and 'Draw
over other apps' permissions when they're not enabled, instead of just
auto-exiting the app since people thought this was the app crashing.
Also improved the Discourage and procedural phrase generation functions
by enabling frequencies of pre/pro words to drop to 0 via Discouraging,
but not allowing the generation algorithm to use them when they're at 0.
The frequencies can still be raised via encouraging or responding.